API: Move mps to top level directory
